Move Parameter and Parameters to gammapy.utils.fitting
