+++ gammapy/utils/fitting/__init__.py
@@ -1,5 +1,5 @@
-"""Fitting utility functions.
-"""
+"""Fitting framework and backends."""
+from .parameter import *
+from .fit import *
 from .iminuit import *
 from .sherpa import *
-from .base import *
