+++ gammapy/utils/fitting/tests/test_parameter.py
@@ -3,8 +3,8 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 import pytest
 import numpy as np
 from numpy.testing import assert_allclose
-from ...extern import six
-from ..modeling import Parameter, Parameters
+from ....extern import six
+from .. import Parameter, Parameters
 
 
 def test_parameter_init():
